chore: divided user into auth and user for scalability
