+++ src/config/authOptions.js
@@ -1,0 +1,5 @@
+export const cookieOptions = {
+  httpOnly: true,
+  sameSite: "none",
+  secure: process.env.NODE_ENV === "production",
+};
